fix(remake header): refactor changes split for different components
Header need to implement language swithcher
