--- src/components/main/header/data.ts
@@ -1,13 +1,0 @@
-export const occupations = {
-  design: ['Figma', 'UI/UX', 'Canva', 'Adobe Illustrator'],
-  front_end: ['React', 'Vue', 'Angular', 'Svelte'],
-  back_end: ['Nest', 'Django', 'Express', 'Laravel'],
-  full_stack: [
-    'React/Node',
-    'Vue/PHP',
-    'Angular/C#',
-    'Svelte/Ruby',
-  ],
-  qa_manual: ['Postman', 'Insomnia'],
-  pm: ['Docker', 'CI/CD', 'Trello', 'Jira'],
-};
